+++ exonum/src/runtime/dispatcher/tests.rs
@@ -94,10 +94,9 @@ impl Dispatcher {
         fork: &Fork,
         artifact: ArtifactId,
         payload: impl BinaryValue,
-    ) -> Result<(), ExecutionError> {
-        Self::commit_artifact(fork, artifact.clone(), payload.to_bytes())?;
+    ) {
+        Self::commit_artifact(fork, artifact.clone(), payload.to_bytes());
         self.block_until_deployed(artifact, payload.into_bytes());
-        Ok(())
     }
 }
 
@@ -216,7 +215,7 @@ impl Runtime for SampleRuntime {
         _snapshot: &dyn Snapshot,
         spec: &InstanceSpec,
         new_status: InstanceStatus,
-    ) -> Result<(), ExecutionError> {
+    ) {
         if spec.artifact.runtime_id == self.runtime_type {
             let status_changed = if let Some(status) = self.services.get(&spec.id).copied() {
                 status != new_status
@@ -228,9 +227,8 @@ impl Runtime for SampleRuntime {
                 self.services.insert(spec.id, new_status);
                 self.new_services.insert(spec.id, new_status);
             }
-            Ok(())
         } else {
-            Err(DispatcherError::IncorrectRuntime.into())
+            panic!(DispatcherError::IncorrectRuntime)
         }
     }
 
@@ -338,12 +336,8 @@ fn test_dispatcher_simple() {
 
     // Check if the services are ready for deploy.
     let mut fork = db.fork();
-    dispatcher
-        .commit_artifact_sync(&fork, rust_artifact.clone(), vec![])
-        .unwrap();
-    dispatcher
-        .commit_artifact_sync(&fork, java_artifact.clone(), vec![])
-        .unwrap();
+    dispatcher.commit_artifact_sync(&fork, rust_artifact.clone(), vec![]);
+    dispatcher.commit_artifact_sync(&fork, java_artifact.clone(), vec![]);
 
     // Check if the services are ready for initiation. Note that the artifacts are pending at this
     // point.
@@ -467,7 +461,7 @@ fn test_dispatcher_simple() {
         .with_runtime(runtime_a.runtime_type, runtime_a)
         .with_runtime(runtime_b.runtime_type, runtime_b)
         .finalize(&blockchain);
-    dispatcher.restore_state(&db.snapshot()).unwrap();
+    dispatcher.restore_state(&db.snapshot());
 
     assert_eq!(
         expected_new_services,
@@ -513,8 +507,7 @@ impl Runtime for ShutdownRuntime {
         _snapshot: &dyn Snapshot,
         _spec: &InstanceSpec,
         _status: InstanceStatus,
-    ) -> Result<(), ExecutionError> {
-        Ok(())
+    ) {
     }
 
     fn execute(
@@ -696,8 +689,7 @@ impl Runtime for DeploymentRuntime {
         _snapshot: &dyn Snapshot,
         _spec: &InstanceSpec,
         _status: InstanceStatus,
-    ) -> Result<(), ExecutionError> {
-        Ok(())
+    ) {
     }
 
     fn execute(
@@ -758,7 +750,7 @@ fn delayed_deployment() {
     // Check that we don't require the runtime to deploy the artifact again if we mark it
     // as committed.
     let fork = db.fork();
-    Dispatcher::commit_artifact(&fork, artifact.clone(), spec).unwrap();
+    Dispatcher::commit_artifact(&fork, artifact.clone(), spec);
     let patch = dispatcher.commit_block_and_notify_runtimes(fork);
     db.merge_sync(patch).unwrap();
     assert_eq!(runtime.deploy_attempts(&artifact), 1);
@@ -784,7 +776,7 @@ fn test_failed_deployment(db: Arc<TemporaryDB>, runtime: DeploymentRuntime, arti
     assert_eq!(runtime.deploy_attempts(&artifact), 1);
 
     let fork = db.fork();
-    Dispatcher::commit_artifact(&fork, artifact, spec).unwrap();
+    Dispatcher::commit_artifact(&fork, artifact, spec);
     dispatcher.commit_block_and_notify_runtimes(fork); // << should panic
 }
 
@@ -828,7 +820,7 @@ fn failed_deployment_with_node_restart() {
     LittleEndian::write_u64(&mut spec, 100);
 
     let fork = db.fork();
-    Dispatcher::commit_artifact(&fork, artifact.clone(), spec).unwrap();
+    Dispatcher::commit_artifact(&fork, artifact.clone(), spec);
     dispatcher.activate_pending(&fork);
     let patch = dispatcher.commit_block_and_notify_runtimes(fork);
     db.merge_sync(patch).unwrap();
@@ -863,7 +855,7 @@ fn recoverable_error_during_deployment() {
     assert_eq!(runtime.deploy_attempts(&artifact), 1);
 
     let fork = db.fork();
-    Dispatcher::commit_artifact(&fork, artifact.clone(), spec).unwrap();
+    Dispatcher::commit_artifact(&fork, artifact.clone(), spec);
     dispatcher.commit_block_and_notify_runtimes(fork);
     // The dispatcher should try to deploy the artifact again despite a previous failure.
     assert!(dispatcher.is_artifact_deployed(&artifact));
@@ -905,9 +897,7 @@ fn stopped_service_workflow() {
         name: "first".into(),
         version: Version::new(0, 1, 0),
     };
-    dispatcher
-        .commit_artifact_sync(&fork, artifact.clone(), vec![])
-        .unwrap();
+    dispatcher.commit_artifact_sync(&fork, artifact.clone(), vec![]);
 
     let service = InstanceSpec {
         artifact: artifact.clone(),
@@ -984,7 +974,7 @@ fn stopped_service_workflow() {
     let mut dispatcher = DispatcherBuilder::new()
         .with_runtime(runtime.runtime_type, runtime)
         .finalize(&blockchain);
-    dispatcher.restore_state(&db.snapshot()).unwrap();
+    dispatcher.restore_state(&db.snapshot());
 
     // Check expected notifications.
     let expected_notifications = vec![